+++ frontend/src/components/schedule/index.ts
@@ -1,0 +1,2 @@
+export { ScheduleListView } from './ScheduleListView'
+export { ScheduleFormView } from './ScheduleFormView'
